requirements-dev.txt
# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.96.1
moto==5.0.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# AWS CDK
aws-cdk-lib==2.120.0
constructs==10.3.0
